Move "parse-name" into the `utils` package
