+++ .changeset/nine-humans-grow.md
@@ -1,0 +1,9 @@
+---
+"@workflow/web-shared": patch
+"@workflow/utils": patch
+"@workflow/core": patch
+"@workflow/cli": patch
+"@workflow/web": patch
+---
+
+Move "parse-name" into the `utils` package
