+++ packages/core/src/runtime.ts
@@ -3,6 +3,7 @@ import {
   WorkflowRunFailedError,
   WorkflowRunNotCompletedError,
 } from '@workflow/errors';
+import { parseWorkflowName } from '@workflow/utils/parse-name';
 import {
   type Event,
   WorkflowInvokePayloadSchema,
@@ -12,7 +13,6 @@ import {
 } from '@workflow/world';
 import { WorkflowSuspension } from './global.js';
 import { runtimeLogger } from './logger.js';
-import { parseWorkflowName } from './parse-name.js';
 import {
   getAllWorkflowRunEvents,
   getQueueOverhead,
